requirements.txt
matplotlib==3.10.7
numpy==2.3.5
opencv_contrib_python==4.10.0.84
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
paddleocr==3.3.2
paddlepaddle_gpu==3.0.0
pandas==2.3.3
Pillow==12.0.0
pytesseract==0.3.13
timm==1.0.22
torch==2.7.1+cu118
tqdm==4.67.1
transformers==4.57.1
ultralytics==8.3.221
